update headers for cmake

+++ src/network/tests/buffer_test.cc
@@ -1,4 +1,4 @@
-#include "../buffer.h"
+#include "network/buffer.h"
 
 #include <sys/types.h>
 #include <sys/stat.h>
@@ -21,23 +21,23 @@ int main() {
   const std::string const_str = "1234567890";
 
   glue_network::ByteBuffer buf;
-  LOG_CHECK(buf.ReadableBytes() == 0, "intiate readable size is non-zero");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == 0, "intiate readable size is non-zero");
   LOG_CHECK(buf.WritableBytes() > 0, "intiate writable size is not BUF_INIT_CAPACITY");
 	
   size += size1;
   buf.Append(str1, size1);
-  LOG_CHECK(buf.ReadableBytes() == size, "Buffer Append error");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == size, "Buffer Append error");
   LOG_INFO("Buf content: %s Append() [PASS] size = %d", buf.ToString().c_str(), buf.ToString().size());
 
   size += size2;
   std::vector<char> char_vec(str2, str2 + size2);
   buf.AppendArray(char_vec);
-  LOG_CHECK(buf.ReadableBytes() == size, "Buffer AppendArray error");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == size, "Buffer AppendArray error");
   LOG_INFO("Buf content: %s AppendArray() [PASS] size = %d", buf.ToString().c_str(), buf.ToString().size());
 
-  size += const_str.size();
+  size += static_cast<int>(const_str.size());
   buf.AppendString(const_str);
-  LOG_CHECK(buf.ReadableBytes() == size, "Buffer AppendString error");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == size, "Buffer AppendString error");
   LOG_INFO("Buf content: %s AppendString() [PASS]", buf.ToString().c_str());
 
   glue_network::ByteBuffer buf2;
@@ -46,13 +46,13 @@ int main() {
   
   size += size3;
   buf.AppendBuffer(buf2);
-  LOG_CHECK(buf.ReadableBytes() == size, "");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == size, "");
   LOG_INFO("Buf content: %s AppendBuffer() [PASS]", buf.ToString().c_str());
 
   size -= size3;
   std::vector<char> char_vec_2 = buf.Read(size3);
   LOG_INFO("Read into vec content: %s ", std::string(&(*char_vec_2.begin()), char_vec_2.size()).c_str());
-  LOG_CHECK(buf.ReadableBytes() == size, "");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == size, "");
   LOG_INFO("Buf content: %s Read() [PASS]", buf.ToString().c_str());
 
   const char* last_pos = buf.FindLast("is");
@@ -69,15 +69,15 @@ int main() {
   umask(0);
   int fd = ::open("buffer_test_txt", O_RDWR | O_CREAT, 0666); /* Permission is needed when creating a new file. */
   int write_cnt = buf.WriteFd(fd);
-  LOG_CHECK(write_cnt == before_str.size(), "WriteFd error");
-  LOG_CHECK(buf.ReadableBytes() == 0, "WriteFd error");
+  LOG_CHECK(write_cnt == static_cast<int>(before_str.size()), "WriteFd error");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == 0, "WriteFd error");
   ::lseek(fd, 0, SEEK_SET); // Need this to seek the file offset to start
   int read_cnt = buf.ReadFd(fd);
   LOG_CHECK(write_cnt == read_cnt, "ReadFd error");
   LOG_INFO("Buf content: %s WriteFd() ReadFd() [PASS]", buf.ToString().c_str());
   
   buf.Reset();
-  LOG_CHECK(buf.ReadableBytes() == 0, "");
+  LOG_CHECK(static_cast<int>(buf.ReadableBytes()) == 0, "");
   LOG_CHECK(buf.AddrOfWrite() == buf.AddrOfRead(), "");
   LOG_INFO("Buf content: %s Reset() [PASS]", buf.ToString().c_str());
